Create Nonlinear Probability of Survival Bins

This function generates nonlinear bins for probability of survival data based on specified thresholds and divisors. It calculates bin statistics, including mean, standard deviation, count, and percentage for each bin.

Usage

nonlinear_bins(
  data,
  Ps_col,
  divisor1 = 5,
  divisor2 = 5,
  threshold_1 = 0.9,
  threshold_2 = 0.99
)

Arguments

data

A `data.frame` or `tibble` containing the probability of survival data for a set of patients.

Ps_col

The column in `data` containing the probability of survival values for a set of patients.

divisor1

A parameter to control the width of the probability of survival range bins. Affects the creation of step sizes for the beginning of each bin range. Defaults to `5`.

divisor2

A parameter to control the width of the probability of survival range bins. Affects the creation of step sizes for the beginning of each bin range. Defaults to `5`.

threshold_1

A parameter to decide where data indices will begin to create step sizes. Defaults to `0.9`.

threshold_2

A parameter to decide where data indices will end to create step sizes. Defaults to `0.99`.

Value

A list with `intervals` and `bin_stats` objects: * `intervals`: A vector of start and end-points for the probability of survival bin ranges. * `bin_stats`: A `tibble` with columns `bin_number`, `bin_start`, `bin_end`, `mean`, `sd`, `count`, and `percent`.

Author

Nicolas Foss, Ed.D, MS, original paper and code in MATLAB by Nicholas J. Napoli, Ph.D., MS

Examples

# Generate example data with high negative skewness
set.seed(123)

# Parameters
n_patients <- 10000  # Total number of patients

# Skewed towards higher values
Ps <- plogis(rnorm(n_patients, mean = 2, sd = 1.5))

# Simulate survival outcomes based on Ps
survival_outcomes <- rbinom(n_patients,
                            size = 1,
                            prob = Ps
                            )

# Create data frame
data <- data.frame(Ps = Ps, survival = survival_outcomes) |>
dplyr::mutate(death = dplyr::if_else(survival == 1, 0, 1))

# Apply the nonlinear_bins function
results <- nonlinear_bins(data = example_data,
                          Ps_col = Ps,
                          divisor1 = 5,
                          divisor2 = 5,
                          threshold_1 = 0.9,
                          threshold_2 = 0.99)
#> Error: object 'example_data' not found

# View results
results$intervals
#> Error: object 'results' not found
results$bin_stats
#> Error: object 'results' not found