Create Nonlinear Probability of Survival Bins

This function generates nonlinear bins for probability of survival data based on specified thresholds and divisors. It calculates bin statistics, including mean, standard deviation, count, and percentage for each bin.

Usage

nonlinear_bins(
  data,
  Ps_col,
  divisor1 = 5,
  divisor2 = 5,
  threshold_1 = 0.9,
  threshold_2 = 0.99
)

Arguments

data

A `data.frame` or `tibble` containing the probability of survival data for a set of patients.

Ps_col

The column in `data` containing the probability of survival values for a set of patients.

divisor1

A parameter to control the width of the probability of survival range bins. Affects the creation of step sizes for the beginning of each bin range. Defaults to `5`.

divisor2

A parameter to control the width of the probability of survival range bins. Affects the creation of step sizes for the beginning of each bin range. Defaults to `5`.

threshold_1

A parameter to decide where data indices will begin to create step sizes. Defaults to `0.9`.

threshold_2

A parameter to decide where data indices will end to create step sizes. Defaults to `0.99`.

Value

A list with `intervals` and `bin_stats` objects: * `intervals`: A vector of start and end-points for the probability of survival bin ranges. * `bin_stats`: A `tibble` with columns `bin_number`, `bin_start`, `bin_end`, `mean`, `sd`, `count`, and `percent`.

Author

Nicolas Foss, Ed.D, MS, original paper and code in MATLAB by Nicholas J. Napoli, Ph.D., MS

Examples

# Generate example data with high negative skewness
set.seed(123)
survival_probs <- rbeta(10000, shape1 = 5, shape2 = 1)
example_data <- data.frame(patient_id = 1:10000, survival_probability = survival_probs)

# Apply the nonlinear_bins function
results <- nonlinear_bins(data = example_data,
                          Ps_col = survival_probability,
                          divisor1 = 5,
                          divisor2 = 5,
                          threshold_1 = 0.9,
                          threshold_2 = 0.99)

# View results
results$intervals
#>  [1] 0.1343271 0.6559647 0.7508905 0.8122535 0.8600744 0.9000012 0.9215288
#>  [8] 0.9412835 0.9585171 0.9743110 0.9899946 0.9999881
results$bin_stats
#> # A tibble: 11 × 7
#>    bin_number bin_start bin_end  mean      sd count percent
#>         <int>     <dbl>   <dbl> <dbl>   <dbl> <dbl>   <dbl>
#>  1          1     0.134   0.656 0.548 0.0910   1181   0.118
#>  2          2     0.656   0.751 0.708 0.0274   1181   0.118
#>  3          3     0.751   0.812 0.784 0.0176   1181   0.118
#>  4          4     0.812   0.860 0.837 0.0138   1181   0.118
#>  5          5     0.860   0.900 0.881 0.0114   1181   0.118
#>  6          6     0.900   0.922 0.911 0.00609   723   0.072
#>  7          7     0.922   0.941 0.932 0.00560   723   0.072
#>  8          8     0.941   0.959 0.950 0.00512   723   0.072
#>  9          9     0.959   0.974 0.967 0.00451   723   0.072
#> 10         10     0.974   0.990 0.982 0.00459   723   0.072
#> 11         11     0.990   1.00  0.995 0.00294   480   0.048